Corrections and Addition of Stylesheet
In this update, we made several significant corrections. Functions related to page configuration have been moved to a new class dedicated to page settings, allowing the Template class to focus solely on the arrangement of elements on the page. Additionally, we added new classes to the stylesheet to define the grid styles that we will use in our templates.